ch
Refactor build-deploy-prod.yml to simplify workflow

The commit simplifies the build-deploy-prod workflow by removing the
unnecessary installation of pnpm and replacing it with npm. This change
reduces complexity and improves maintainability. Additionally, the
commit ensures that the workflow is triggered only on push events to the
master branch, streamlining the deployment process.

#ch